requirements.txt
flask>=2.3.0
google-cloud-firestore>=2.11.0
google-cloud-functions>=1.20.0
google-cloud-run>=0.10.0
google-cloud-pubsub>=2.18.0
google-cloud-monitoring>=2.16.0
google-cloud-logging>=3.8.0
pydantic>=2.4.0
openai>=1.3.0
algoliasearch==3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
requests>=2.32.3
aiohttp>=3.10.11
pytest>=7.4.0
pytest-flask>=1.2.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0
bcrypt
PyJWT
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
spotipy>=2.25.0
scikit-learn>=1.7.0
numpy>=2.3.0
pandas>=2.3.0 